CLN: consolidate Series.quantile and DataFrame.quantile
